+++ docs/changelog.md
@@ -60,6 +60,7 @@ As features stabilize some brief notes about them will accumulate here.
   [wezterm cli set-window-title](cli/cli/set-window-title.md). #522 #1598
 * [wezterm cli rename-workspace](cli/cli/rename-workspace.md). #2787
 * [wezterm.mux.rename_workspace](config/lua/wezterm.mux/rename_workspace.md). #2787
+* [wezterm cli get-pane-direction](cli/cli/get-pane-direction.md)
 
 #### Fixed
 * mux: Stale remote window mapping could prevent spawning new tabs in remote domain. #2759